+++ ci/clang_format.py
@@ -86,7 +86,6 @@ EXCLUDE = (
     "cmd/tools/cvtgxl.c",
     "cmd/tools/dijkstra.c",
     "cmd/tools/gc.c",
-    "cmd/tools/gml2gv.c",
     "cmd/tools/gml2gv.h",
     "cmd/tools/graph_generator.c",
     "cmd/tools/graphml2gv.c",